Update send_to_discord function

+++ utils/handle_send_to_discord.py
@@ -3,17 +3,20 @@ from utils.store_conversation import store_conversation
 import queue as thread_queue
 import threading
 import asyncio
+from openai import OpenAI
 import openai
 from utils.exponential_backoff import exponential_backoff
 import concurrent.futures
 
 def threaded_fetch(response, queue, completion):
     for chunk in response:
-        new_content = chunk.choices[0].delta.get("content", "")
+        # print(chunk)
+        new_content = chunk.choices[0].delta.content if chunk.choices[0].delta.content is not None else ""
+        # print(new_content)
         if new_content:
             completion += new_content
             queue.put(new_content)
-    print("Response generated from GPT-4!")
+    print("Response generated!")
     queue.put(None)
     return completion
 
@@ -48,14 +51,21 @@ async def send_to_discord(queue, min_chunk_size, max_chunk_size, delay, temp_mes
 
         # Send the first chunk after 0.3 seconds regardless of its size
         if is_first_chunk:
+            if full_response == '':
+                continue
             print("started to send the first chunk")
-            await asyncio.sleep(0.1)
+            # await asyncio.sleep(0.1)
             current_content = full_response
             full_response = ""  # Clear out full_response since we've processed its content
             if temp_message:
                 await temp_message.delete()
             try:
-                temp_message = await message.channel.send(current_content)
+                if current_content != '':
+                    temp_message = await message.channel.send(current_content)
+                    
+                else:
+                    temp_message = await message.channel.send(".")
+                    
             except discord.errors.HTTPException:
                 embed = create_embed(current_content)
                 embed_message = await message.channel.send(embed=embed)
@@ -64,9 +74,9 @@ async def send_to_discord(queue, min_chunk_size, max_chunk_size, delay, temp_mes
             continue  # Skip to the next iteration to check for more content
         
         if i == 1:
-            min_chunk = 10
+            min_chunk = 20
         else:
-            min_chunk = 50
+            min_chunk = 75
         
         # Process other chunks when size >= 50
         while len(full_response) >= min_chunk:
@@ -81,8 +91,9 @@ async def send_to_discord(queue, min_chunk_size, max_chunk_size, delay, temp_mes
                     await embed_message.edit(embed=embed)
                 else:
                     temp_message = await temp_message.channel.fetch_message(temp_message.id)
-                    await temp_message.edit(content=temp_message.content + current_content)
-            except discord.errors.HTTPException:
+                    await temp_message.edit(final_response)
+            except discord.errors.HTTPException as e:
+                print(e)
                 if temp_message:
                     await temp_message.delete()  # <-- Delete temp_message if it exists
                 if not embed_message:
@@ -115,20 +126,21 @@ async def send_to_discord(queue, min_chunk_size, max_chunk_size, delay, temp_mes
 
 
 # Convert the lambda to a regular function
-def synchronous_generate_response(model, latest_conversation):
-    return openai.ChatCompletion.create(
+def synchronous_generate_response(model, latest_conversation, client):
+    return client.chat.completions.create(
         model=model,
         messages=latest_conversation,
         stream=True,
-        allow_fallback=True
+        allow_fallback=True,
+        # premium=True
     )
 
-async def generate_response(conversation, message, conversation_id):
+async def generate_response(conversation, message, conversation_id, client):
     loop = asyncio.get_running_loop()
 
     # Run the synchronous call inside a thread
     with concurrent.futures.ThreadPoolExecutor() as pool:
-        api_call = lambda model, latest_conversation: loop.run_in_executor(pool, synchronous_generate_response, model, latest_conversation)
+        api_call = lambda model, latest_conversation: loop.run_in_executor(pool, synchronous_generate_response, model, latest_conversation, client)
         
         response = await exponential_backoff(
             api_call,
@@ -138,7 +150,7 @@ async def generate_response(conversation, message, conversation_id):
 
     return response
 
-async def update_conversation_and_send_to_discord(function_response, function_name, temp_message, conversation, conversation_id, message):
+async def update_conversation_and_send_to_discord(function_response, function_name, temp_message, conversation, conversation_id, message, client):
     conversation.append(
         {
             "role": "function",
@@ -150,7 +162,7 @@ async def update_conversation_and_send_to_discord(function_response, function_na
 
     final_response = ""
     try:
-        response = await generate_response(conversation, message, conversation_id)
+        response = await generate_response(conversation, message, conversation_id, client)
     except Exception as e:
         print(f"Error occurred: {e}")
         temp_message.delete()
@@ -159,7 +171,7 @@ async def update_conversation_and_send_to_discord(function_response, function_na
     thread_safe_queue = thread_queue.Queue()
     threading.Thread(target=threaded_fetch, args=(response, thread_safe_queue, "")).start()
 
-    completion, temp_message = await send_to_discord(thread_safe_queue, 75, 2000, 0.25, temp_message, final_response, message)
+    completion, temp_message = await send_to_discord(thread_safe_queue, 75, 2000, 0.3, temp_message, final_response, message)
 
     conversation.append(
         {
